Abandoning all professionalism and invoking the dark gods (using an LLM)
I was getting a bit of help before, but at this point I'm just going to throw a bunch of shit at the wall and see what sticks. In my opinion, using LLMs is a crutch and should not be used in a professional environment at all, but for the case of a workflow file, I think it's acceptable. Besides, once I submit something working, someone actually smart can look at it and make it better.