Revert; Exhausted my last idea
That might be curtains, I don't know what else to do.

Post mortem: I managed to get it to the point where it outputs the .app file, but basically every library is missing in spite of the shell script I started with including them just fine. Copying them into the framework folder from a build made by the shell script would make it launch just fine, IF it wasn't for the fact that MacOS marks the build as broken because it doesn't like the way it's signed. This is at least something I kind of know how to properly fix, but I think I'd need to upload my Apple ID credentials to Github here, and I'm pretty sure that would have to be public... so that's not viable for what I'm trying to do here.

All in all, I did my best, I might even try a few more things tomorrow, but I'm pretty sure this is as far as I'm getting.